TKH-3141 Correctly handle missing values in lists and maps
This can lead to missing values all the way down, so we need to handle null-safe conversion all the way down.
We only know whether we need to use the value from plan or config values when calling the setter, when setting the actual attribute, so until then we need to fake an object graph with all attributes and values present (even if as null).