TKH-3142 Add and use SDKInterfaceTypeName where appropriate
The SDKTypeName methode really returned the interface type, not the concrete type.
So we renamed that method, including all usages, to
'SDKInterfaceTypeName' so the naming fits betteri.
We also need the actual type in a few spots in which we now akwardly recreate it based on the wrong property which doesn't fit as well.
So we added an actual 'SDKTypeName' method that returns the actual type
and use that.